CryptStream/pyproject.toml
[project]
name = "cryptstream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
